Fix frontend error out
