+++ src/commands/list.ts
@@ -5,7 +5,10 @@ import log, { LevelNames } from '../utils/log';
 import AWSWrapper from '../aws';
 import ora from 'ora';
 import { ensureCredentials } from '../utils/aws';
-import { logRoles, logBuckets, logDeliveryStreams, logLambdas, logTimestream, logTimestreamTables } from '../utils/log';
+import chalk from 'chalk';
+import cli from 'cli-ux';
+
+import { ALL_TIMESTREAM_DATABASE_TABLES } from '../constants';
 
 export default class ListCommand extends Command {
   static description = 'describe the command here';
@@ -42,6 +45,110 @@ Timestream
 
   static args = [];
 
+  async logRoles(roles: { RoleName: string}[], callback: (msg: string) => void): Promise<void> {
+    callback(chalk.bold('Role:'));
+
+    if (roles.length === 0) {
+      log.info('No role found!');
+      return;
+    }
+
+    for await (const role of roles) {
+      cli.url(
+        `- ${role.RoleName}`,
+        `https://console.aws.amazon.com/iam/home?region=us-east-1#/roles/${role.RoleName}`
+      );
+    }
+
+    console.log('\n');
+  }
+
+  async logBuckets(buckets: { Name: string}[], callback: (msg: string) => void): Promise<void> {
+    callback(chalk.bold('Bucket:'));
+
+    if (buckets.length === 0) {
+      log.info('No bucket found!');
+      return;
+    }
+
+    for (const bucket of buckets) {
+      cli.url(
+        `- ${bucket.Name}`,
+        `https://s3.console.aws.amazon.com/s3/buckets/${bucket.Name}?region=us-east-2&tab=objects`
+      );
+    }
+    console.log('\n');
+  }
+
+  async logDeliveryStreams(streams: string[], callback: (msg: string) => void): Promise<void> {
+    callback(chalk.bold('Firehose Stream:'));
+
+    if (streams.length === 0) {
+      log.info('No stream found!');
+      return;
+    }
+
+    for await (const stream of streams) {
+      cli.url(`- ${stream}`, `https://console.aws.amazon.com/firehose/home?region=us-east-1#/details/${stream}`);
+    }
+    console.log('\n');
+  }
+
+  async logLambdas(lambdas: { FunctionName: string }[], callback: (msg: string) => void): Promise<void> {
+    callback(chalk.bold('Lambda:'));
+
+    if (lambdas.length === 0) {
+      log.info('No lambda found!');
+      return;
+    }
+
+    for await (const lambda of lambdas) {
+      cli.url(
+        `- ${lambda.FunctionName}`,
+        `https://console.aws.amazon.com/lambda/home?region=us-east-1#/functions/${lambda.FunctionName}?tab=code`
+      );
+    }
+
+    console.log('\n');
+  }
+
+  async logTimestream(streams: { DatabaseName: string }[], callback: (msg: string) => void): Promise<void> {
+    callback(chalk.bold('Timestream Database:'));
+
+    if (streams.length === 0) {
+      log.info('No timestream database found!');
+      return;
+    }
+
+    for await (const stream of streams) {
+      cli.url(
+        `- ${stream.DatabaseName}`,
+        `https://console.aws.amazon.com/timestream/home?region=us-east-1#databases/${stream.DatabaseName}`
+      );
+    }
+    console.log('\n');
+  }
+
+  async logTimestreamTables(tables: { TableName: string, DatabaseName: string }[], callback: (msg: string) => void): Promise<void> {
+    callback(chalk.bold('Timestream Tables:'));
+
+    if (!tables || tables.length === 0) {
+      log.info('No timestream tables found!');
+      return;
+    }
+
+    for await (const table of tables) {
+      if (!ALL_TIMESTREAM_DATABASE_TABLES.includes(table.TableName)) {
+        continue;
+      }
+
+      cli.url(
+        `- ${table.TableName}`,
+        `https://console.aws.amazon.com/timestream/home?region=us-east-1#databases/DendroTimestreamDB/tables/${table.TableName}`
+      );
+    }
+  }
+
   async run(): Promise<void>{
     ensureCredentials();
 
@@ -53,28 +160,28 @@ Timestream
 
     spinner = log.spin('Listing role...');
     const Roles = await AWSWrapper.listRoles();
-    await logRoles(Roles, callback);
+    await this.logRoles(Roles, callback);
 
     spinner = log.spin('Listing S3 bucket...');
     const Buckets = await AWSWrapper.listBuckets();
-    await logBuckets(Buckets, callback);
+    await this.logBuckets(Buckets, callback);
 
     spinner = log.spin('Listing Firehose stream...\n');
     const DeliveryStreamNames = await AWSWrapper.listDeliveryStreams();
-    await logDeliveryStreams(DeliveryStreamNames, callback);
+    await this.logDeliveryStreams(DeliveryStreamNames, callback);
 
     spinner = log.spin('Listing Lambda function...\n');
     const Functions = await AWSWrapper.listFunctions();
-    await logLambdas(Functions, callback);
+    await this.logLambdas(Functions, callback);
 
     spinner = log.spin('Listing Timestream database...\n');
     const Databases = await AWSWrapper.listDatabases();
-    await logTimestream(Databases, callback);
+    await this.logTimestream(Databases, callback);
 
     if ( Databases.length > 0 ) {
       spinner = log.spin('Listing Timestream tables...\n');
       const Tables = await AWSWrapper.listTables();
-      await logTimestreamTables(Tables, callback);
+      await this.logTimestreamTables(Tables, callback);
     }
   }
 }
